Refactor: Split/arrange exception factories
