+++ src/flatcollections-array/FlatArray/FlatArray.T.Builder/FlatArray.Builder.EnsureCapacity.cs
@@ -9,7 +9,7 @@ partial struct FlatArray<T>
         {
             if (capacity is not >= 0)
             {
-                throw InnerExceptionFactory.CapacityOutOfRange_MustBeGreaterThanOrEqualToZero(nameof(capacity), capacity);
+                throw InnerBuilderExceptionFactory.CapacityOutOfRange_MustBeGreaterThanOrEqualToZero(nameof(capacity), capacity);
             }
 
             // Copy the state to reduce the chance of multithreading side effects
