+++ src/flatcollections-array/FlatArray/FlatArray.T.InnerFlatList/FlatArray.InnerFlatList.Impl.cs
@@ -26,7 +26,7 @@ partial struct FlatArray<T>
             }
             set
             {
-                throw InnerExceptionFactory.NotSupportedOnReadOnlyArray();
+                throw InnerListExceptionFactory.NotSupportedOnReadOnlyArray();
             }
         }
