+++ src/flatcollections-array/FlatArray/FlatArray.T.InnerFlatList/InnerListExceptionFactory/FlatArray.InnerFlatList.InnerListExceptionFactory.cs
@@ -1,0 +1,14 @@
+﻿namespace System;
+
+partial struct FlatArray<T>
+{
+    partial class InnerFlatList
+    {
+        private static class InnerListExceptionFactory
+        {
+            internal static NotSupportedException NotSupportedOnReadOnlyArray()
+                =>
+                new("The operation is not supported on read-only array.");
+        }
+    }
+}
