+++ src/flatcollections-array/FlatArray/FlatArray.T.JsonConverter/InnerJsonExceptionFactory/FlatArray.JsonConverter.InnerJsonExceptionFactory.cs
